+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Update base URL to reflect the changes in Robokassa API.
 * Allow `pay_url` to accept block
 * Hash algorithms implemented (Алгоритм расчёта хэша — http://docs.robokassa.ru/#3733) [#29][]
+* Receipt handling implemented (12/2023).
 
 [#23]:https://github.com/ZeroOneStudio/rubykassa/pull/23
 [#24]:https://github.com/ZeroOneStudio/rubykassa/pull/24